requirements.txt
flask==2.3.3
requests==2.31.0
PySocks==1.7.1
urllib3==2.0.7
